requirements.txt
requests
nlpcloud
python-dotenv
matplotlib
numpy
scipy
pandas
praat-parselmouth
